Grouped a few rules (tag stuff and entities) into one repository rule (which can then be included).
I didn't mention this in the release notes, but it is actually possible to let a repository rule group several rules, the syntax is a normal rule without begin/end patterns, but a patterns array.

I didn't mention it because the patterns key is there only to ease the implementation. It does however (sort of) make sense, since an empty begin pattern means the rule matches everything (thus the sub-patterns will be applied), and the same goes for an empty end pattern (so if none of the sub-patterns match, it'll leave the sub-patterns again) -- and it could be useful to name a group in the future (which would then be a parent name to all sub-patterns, which may either be because they all are anyway named the same, or because we want to make an “alias” for a rule (although this can already be done via captures)).

So all in all I'm leaving it as-is, also because I hope the final pieces are soon falling into place wrt the graphical editor for language grammars! :)



git-svn-id: http://svn.textmate.org/trunk/Bundles/HTML.tmbundle@1372 dfb7d73b-c2ec-0310-8fea-fb051d288c6d