• replaced the old link-macro with something I think is of more general use.
It wraps the selection in <a href="...">«selection»</a> but makes the link be the content of the clipboard (if the clipboard doesn't contain linefeeds).

It inserts the text as a snippet with the link selected and a tab-stop past the end tag, so simply press tab if the link is as desired (otherwise fix it and then press tab).

The key equivalent for this command is cmd-shift-L (like the old macro).


git-svn-id: http://svn.textmate.org/trunk/Bundles/HTML.tmbundle@1501 dfb7d73b-c2ec-0310-8fea-fb051d288c6d