Added 4 rules to classify a tag as either:
  · entity.name.tag.structure.any -- html, body, head
  · entity.name.tag.block.any     -- what should be block in HTML 4
  · entity.name.tag.inline.any    -- the rest of the valid HTML 4 tags
  · entity.name.tag.other         -- catch-all

I put “any” in the scope to signal that it's a catch-all rule. So e.g. one could specialize it by adding:
  · entity.name.tag.inline.img
And it'd be possible to style all inline tags except those with specific rules (although providing a style for the latter would override a general style for entity.name.tag.inline).

It may make sense to form other groups, I rarely work with HTML and have little preference wrt syntax highlight, so I'll leave it to others to further refine this specialized markup of tag names.

There's a rule in the HTML grammar to catch empty tag-pairs, this one will win over these four (for empty tag-pairs) -- I left it as-is, since the fix I'd prefer requires one (planned) addition to the grammar capabilities. 



git-svn-id: http://svn.textmate.org/trunk/Bundles/HTML.tmbundle@1542 dfb7d73b-c2ec-0310-8fea-fb051d288c6d