Fix exception presumably introduced with Lion.
The checkbox in the nib previously resulted in a string or integer being stored in the property list returned. We used `to_i` and compared it with `1` for “maximum flexibility”.

Now an actual boolean is found in the property list, and as the boolean does not implement the `to_i` method, an exception was thrown.

This fix has not been tested on 10.6.