Only use “current word” if it looks like a tag.
This is done by matching against /^[\w\-:_]+$/i. In practice current word would only contain a dash, underscore, or colon, if the character classes for the document type has been redefined to make that part of the word character class or if the command is invoked with a selection.

This change was made to avoid an issue when using the command in a location like: <p>‸</p>, here “current word” would be “></”.

Also enable UTF-8 so /\w/ matches non-ASCII word characters.