Merge Href drag commands into single command
To simplify this has been merged into a single command, will try to insert a proper amount of whitespace depending on insertion point. Drops the title attribute as while this is desirable if you are inserting into an existing tag we can't know what already exists. The standard anchor drag command continues to include the title attribute.

The addition of `B:` prefix ensures we are inside a tag rather than just beside one.