Add HTML (Derivative) language
This is for cases where you want to use the HTML grammar but without the invalid tag matching. This is most common in template languages or other cases where the text is put through a pre-processor. This does not have any file extensions it is assigned to but can be selected manually.

Fixes #92.